Deep modifications. The model now supports convolutional AND linear data inputs, in a way that at least feels more appropriate than dropping the data onto the board somewhere.
The model was flying blind for some period of time. I was passing in an empty copy of the board, for prediction. I committed the chosen action to the real board, but I was holding a reference to an empty board (created at the start of the episode) and kept passing that into the predict() method. IF there's some massive bug that had been holding me back. This may have been it, or one of them.

zero-aligned mino enum, cleanup, and more.